// app/backend/src/interfaces/@csvSeeder.ts
export default interface csvSeeder {
  reference: string;
  date: string;
  checkIn: string;
  checkOut: string;
  rentPrice: string;
  rentWithoutTax: string;
  rentTax: string;
  extras: string;
  extrasWithoutTax: string;
  extrasTax: string;
  total: string;
  totalWithoutTax: string;
  totalTax: string;
  payment: string;
  pending: string;
  property: string;
  locality: string;
  portal: string;
  calculatedCommission: string;
  personalizedCommission: string;
}